feat: update tests for electra
