+++ Cargo.toml
@@ -17,8 +17,9 @@ members = [
     "crates/runtime",
     "crates/storage",
     "testing/beacon-api",
-    "testing/ef-tests",
-]
+    "testing/beacon-api"
+,
+    "testing/ef-tests"]
 resolver = "2"
 exclude = ["book/cli", "book/sources"]
