+++ crates/common/consensus/src/bls_to_execution_change.rs
@@ -4,6 +4,8 @@ use serde::{Deserialize, Serialize};
 use ssz_derive::{Decode, Encode};
 use tree_hash_derive::TreeHash;
 
+use crate::misc::checksummed_address;
+
 #[derive(Debug, PartialEq, Clone, Serialize, Deserialize, Encode, Decode, TreeHash)]
 pub struct SignedBLSToExecutionChange {
     pub message: BLSToExecutionChange,
@@ -15,5 +17,6 @@ pub struct BLSToExecutionChange {
     #[serde(with = "serde_utils::quoted_u64")]
     pub validator_index: u64,
     pub from_bls_pubkey: PubKey,
+    #[serde(with = "checksummed_address")]
     pub to_execution_address: Address,
 }
