+++ crates/common/consensus/src/consolidation_request.rs
@@ -4,8 +4,11 @@ use serde::{Deserialize, Serialize};
 use ssz_derive::{Decode, Encode};
 use tree_hash_derive::TreeHash;
 
+use crate::misc::checksummed_address;
+
 #[derive(Debug, PartialEq, Clone, Serialize, Deserialize, Encode, Decode, TreeHash)]
 pub struct ConsolidationRequest {
+    #[serde(with = "checksummed_address")]
     pub source_address: Address,
     pub source_pubkey: PubKey,
     pub target_pubkey: PubKey,
