+++ crates/common/consensus/src/misc.rs
@@ -15,6 +15,27 @@ use crate::{
     signing_data::SigningData,
 };
 
+pub mod checksummed_address {
+    use alloy_primitives::Address;
+    use serde::{Deserialize, Deserializer, Serializer, de::Error as _};
+
+    pub fn serialize<S>(address: &Address, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        let checksummed = address.to_checksum(None);
+        serializer.serialize_str(&checksummed)
+    }
+
+    pub fn deserialize<'de, D>(deserializer: D) -> Result<Address, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let s: String = Deserialize::deserialize(deserializer)?;
+        s.parse::<Address>().map_err(D::Error::custom)
+    }
+}
+
 pub fn compute_signing_root<SSZObject: TreeHash>(ssz_object: SSZObject, domain: B256) -> B256 {
     SigningData {
         object_root: ssz_object.tree_hash_root(),
