+++ crates/common/consensus/src/withdrawal.rs
@@ -4,6 +4,8 @@ use serde::{Deserialize, Serialize};
 use ssz_derive::{Decode, Encode};
 use tree_hash_derive::TreeHash;
 
+use crate::misc::checksummed_address;
+
 #[derive(
     Debug, PartialEq, Clone, Serialize, Deserialize, Encode, Decode, TreeHash, RlpEncodable,
 )]
@@ -12,6 +14,7 @@ pub struct Withdrawal {
     pub index: u64,
     #[serde(with = "serde_utils::quoted_u64")]
     pub validator_index: u64,
+    #[serde(with = "checksummed_address")]
     pub address: Address,
     #[serde(with = "serde_utils::quoted_u64")]
     pub amount: u64,
